requirements.txt
autopep8==1.6.0
cycler==0.11.0
imageio==2.10.3
kiwisolver==1.3.2
matplotlib==3.4.3
networkx==2.6.3
opencv-python==4.5.4.58
Pillow==8.4.0
pycodestyle==2.8.0
pyparsing==3.0.4
python-dateutil==2.8.2
PyWavelets==1.1.1
scikit-image==0.18.3
scipy==1.7.2
six==1.16.0
tifffile==2021.11.2
toml==0.10.2
